add set-token command and restructure
